feat: merge published and updated dates if identical
In the post description, merge "posted on X and updated on X" if the two dates are the same, instead reading "posted and updated on X". Retain the old text if they are different.

It is necessary to check if page.updated exists as well as if date and updated are different otherwise Zola will attempt to evaluate their equality even if updated doesn't exist, which will lead to an error. Likewise, it is required to also evaluate to true if updated doesn't exist (hence `or not page.updated`), otherwise posts without updated dates won't have the publication date listed.