refactor: mastodon link verification
Merge Mastodon and non-Mastodon links into one `<a>` tag, adding `rel="me"` if the link is a Mastodon link.

The previous solution used two different `<a>` tags, one for Mastodon links and the other for all others. This would require changes made to one to be replicated in the other manually.